Update HelloWorld.sol
Access Control: Added a onlyOwner modifier to restrict the ability to change the greeting to the owner of the contract (the one who deployed it).

Owner Variable: Introduced an owner state variable to keep track of the contract owner.

Transfer Ownership Function: An optional function to change the owner of the contract. This is useful if you ever need to transfer control of the contract to a different address.

Error Handling: Added require statements for validating conditions (like non-zero new owner address).

Event Indexing: Added indexed to the sender in the event for easier filtering and searching.

Signed-off-by: Nealsen <soumitrarocking05@gmail.com>